+++ CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Fixes batch url computation
 
+### Added
+
+- Event aggregation feature
+
 ## [1.6.0] - 2023-10-20
 
 ### Added